requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-import-export==3.3.3
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.9
PyYAML==6.0.1
setuptools==68.2.2
sqlparse==0.4.4
tablib==3.5.0
wheel==0.41.2
xlrd==2.0.1
xlwt==1.3.0
